Refactored panel out of searchPanel/index.js and into own component
