--- spec/models/user_spec.rb
@@ -1,4 +1,0 @@
-
-RSpec.describe User do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
